requirements.txt
requests
python-dotenv
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytest
slack-sdk
slack-bolt
fastapi
uvicorn
ddgs
anthropic
schedule
psutil
mcp
httpx
pylint
